# Remove all reflection for locating defaulting functions