… pkg/api/serialization_proto_test.go
… pkg/api/v1/conversion_generated.go
… pkg/apis/authorization/v1beta1/conversion_generated.go
… pkg/apis/autoscaling/v1/conversion_generated.go
… pkg/apis/batch/v1/conversion_generated.go
… pkg/apis/extensions/v1beta1/conversion.go
… pkg/apis/extensions/v1beta1/conversion_generated.go
… pkg/apis/metrics/v1alpha1/conversion_generated.go
… pkg/apis/policy/v1alpha1/conversion_generated.go