@smarterclayton 2016-04-29
Code is functionally equivalent - we already require conversions to be generated on every change.

@kubernetes/sig-api-machinery