@wojtek-t 2016-04-29
LGTM (just a very minor comment).