@wojtek-t 2016-04-29
@smarterclayton - since the first 2 commits are already merged, can you please remove them?
And then feel free to self-lgtm.