@smarterclayton 2016-04-29
I can't merge this until those are merged.  If those commits are removed this PR breaks.