@smarterclayton 2016-04-29
applying LGTM after addressing comment