@wojtek-t 2016-05-06
@smarterclayton - can you please rebase?