@roberthbailey 2016-05-06
I was merging all of the small PRs that I could find. This one had a much bigger +/- count so I hadn't looked at it. With the manual merges, it's now #2 in the submit queue so it should get automatically merged once the kubemark-500 test starts passing again.