@smarterclayton 2016-05-07
Trivial rebase, reapplying lgtm