@wojtek-t 2016-05-09
@smarterclayton - are you looking into perf-dash.k8s.io ?
If so, what chart specifically are you looking at?