@smarterclayton 2016-05-16
It was on load for node / pods.  But after a few builds it corrected, so I now doubt that the test was capturing the result of the change.